requirements.txt
numpy>=1.11.0
pandas>=0.18.1
autograd
Pillow

